backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyJWT==2.10.1
SQLAlchemy==2.0.41
psycopg[binary]==3.2.9
pytest==8.4.1
httpx==0.28.1
email-validator==2.2.0
